requirements.txt
asgiref==3.8.1
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
dj-database-url==3.0.0
Django==4.2.14
django-storages==1.14.6
dotenv==0.9.9
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
python-magic==0.4.27